LayoutTests/imported/w3c/web-platform-tests/intersection-observer/iframe-no-root-expected.txt



PASS Observer with the implicit root; target in a same-origin iframe. 
FAIL First rAF. assert_equals: entries.length expected 1 but got 0
FAIL document.scrollingElement.scrollTop = 200 assert_equals: entries.length == 1 expected 1 but got 0
FAIL iframe.contentDocument.scrollingElement.scrollTop = 250 assert_equals: entries.length expected 2 but got 0
FAIL document.scrollingElement.scrollTop = 100 assert_equals: entries.length expected 3 but got 0
